fix: fix dark/light theme injection
